+++ tests/Process/StartCondition/CanBeStartedTest.php
@@ -14,9 +14,9 @@ use Steevanb\ParallelProcess\{
 /**
  * @covers \Steevanb\ParallelProcess\Process\StartCondition::__construct
  * @covers \Steevanb\ParallelProcess\Process\StartCondition::canBeStarted
- * @covers \Steevanb\ParallelProcess\Process\StartCondition::addProcessTerminated
- * @covers \Steevanb\ParallelProcess\Process\StartCondition::addProcessSuccessful
- * @covers \Steevanb\ParallelProcess\Process\StartCondition::addProcessFailed
+ * @covers \Steevanb\ParallelProcess\Process\StartCondition::getProcessesTerminated
+ * @covers \Steevanb\ParallelProcess\Process\StartCondition::getProcessesSuccessful
+ * @covers \Steevanb\ParallelProcess\Process\StartCondition::getProcessesFailed
  */
 final class CanBeStartedTest extends TestCase
 {
@@ -33,7 +33,7 @@ final class CanBeStartedTest extends TestCase
     {
         $startCondition = new StartCondition($this->createLsProcess());
         $process = $this->createLsProcess();
-        $startCondition->addProcessTerminated($process);
+        $startCondition->getProcessesTerminated()->add($process);
 
         static::assertFalse($process->isTerminated());
         static::assertFalse($startCondition->canBeStarted());
@@ -44,7 +44,7 @@ final class CanBeStartedTest extends TestCase
         $startCondition = new StartCondition($this->createLsProcess());
         $process = $this->createLsProcess();
         $process->run();
-        $startCondition->addProcessTerminated($process);
+        $startCondition->getProcessesTerminated()->add($process);
 
         static::assertTrue($process->isTerminated());
         static::assertTrue($startCondition->canBeStarted());
@@ -54,7 +54,7 @@ final class CanBeStartedTest extends TestCase
     {
         $startCondition = new StartCondition($this->createLsProcess());
         $process = $this->createLsProcess();
-        $startCondition->addProcessSuccessful($process);
+        $startCondition->getProcessesSuccessful()->add($process);
 
         static::assertFalse($process->isTerminated());
         static::assertFalse($startCondition->canBeStarted());
@@ -65,7 +65,7 @@ final class CanBeStartedTest extends TestCase
         $startCondition = new StartCondition($this->createLsProcess());
         $process = $this->createLsProcess();
         $process->mustRun();
-        $startCondition->addProcessSuccessful($process);
+        $startCondition->getProcessesSuccessful()->add($process);
 
         static::assertTrue($process->isTerminated());
         static::assertTrue($process->isSuccessful());
@@ -76,7 +76,7 @@ final class CanBeStartedTest extends TestCase
     {
         $startCondition = new StartCondition($this->createLsProcess());
         $process = new Process(['unknown-command']);
-        $startCondition->addProcessFailed($process);
+        $startCondition->getProcessesFailed()->add($process);
 
         static::assertFalse($process->isTerminated());
         static::assertFalse($startCondition->canBeStarted());
@@ -87,7 +87,7 @@ final class CanBeStartedTest extends TestCase
         $startCondition = new StartCondition($this->createLsProcess());
         $process = new Process(['unknown-command']);
         $process->run();
-        $startCondition->addProcessFailed($process);
+        $startCondition->getProcessesFailed()->add($process);
 
         static::assertTrue($process->isTerminated());
         static::assertFalse($process->isSuccessful());